+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.Automations/Automation/AutomationService.cs
@@ -24,7 +24,13 @@ namespace hhnl.HomeAssistantNet.Automations.Automation
         /// </summary>
         /// <param name="automation">The automation to start.</param>
         /// <param name="changedEntity">The entity who's changes caused this automation to be enqueued.</param>
-        Task EnqueueAutomationForEntityChangeAsync(AutomationEntry automation, string changedEntity);
+        Task EnqueueAutomationForEntityChangeAsync(AutomationEntry automation, string changedEntity, Event @event);
+
+        /// <summary>
+        /// Enqueues a new automation and waits for it's start.
+        /// </summary>
+        /// <param name="automation">The automation to start.</param>
+        Task EnqueueAutomationForEventFiredAsync(AutomationEntry automation, Event @event);
 
         /// <summary>
         /// Enqueues a new automation and waits for it's start.
@@ -58,16 +64,21 @@ namespace hhnl.HomeAssistantNet.Automations.Automation
             _serviceProvider = serviceProvider;
         }
 
-        public async Task EnqueueAutomationForEntityChangeAsync(AutomationEntry automation, string changedEntity)
+        public async Task EnqueueAutomationForEntityChangeAsync(AutomationEntry automation, string changedEntity, Event @event)
         {
-            await EnqueueAutomationRunAsync(automation, AutomationRunInfo.StartReason.EntityChanged, changedEntity, null);
+            await EnqueueAutomationRunAsync(automation, AutomationRunInfo.StartReason.EntityChanged, changedEntity, null, @event);
+        }
+
+        public async Task EnqueueAutomationForEventFiredAsync(AutomationEntry automation, Event @event)
+        {
+            await EnqueueAutomationRunAsync(automation, AutomationRunInfo.StartReason.EventFired, null, null, @event);
         }
 
         public async Task EnqueueAutomationForManualStartAsync(AutomationEntry automation)
         {
             TaskCompletionSource tcs =
                 new(TaskCreationOptions.RunContinuationsAsynchronously);
-            await EnqueueAutomationRunAsync(automation, AutomationRunInfo.StartReason.Manual, null, tcs);
+            await EnqueueAutomationRunAsync(automation, AutomationRunInfo.StartReason.Manual, null, tcs, Event.Empty);
             await tcs.Task;
         }
 
@@ -94,7 +105,7 @@ namespace hhnl.HomeAssistantNet.Automations.Automation
             // Start all automations that are configure to run on startup
             foreach (AutomationEntry? runOnStartAutomations in _automationRegistry.Automations.Values.Where(a => a.Info.RunOnStart))
             {
-                await EnqueueAutomationRunAsync(runOnStartAutomations, AutomationRunInfo.StartReason.RunOnStart, null, null);
+                await EnqueueAutomationRunAsync(runOnStartAutomations, AutomationRunInfo.StartReason.RunOnStart, null, null, Event.Empty);
             }
 
             // Start schedules
@@ -132,7 +143,7 @@ namespace hhnl.HomeAssistantNet.Automations.Automation
                         return;
                     }
 
-                    await EnqueueAutomationRunAsync(entry, AutomationRunInfo.StartReason.Schedule, null, null);
+                    await EnqueueAutomationRunAsync(entry, AutomationRunInfo.StartReason.Schedule, null, null, Event.Empty);
 
                     ScheduleNextRun(entry);
                 }
@@ -200,7 +211,8 @@ namespace hhnl.HomeAssistantNet.Automations.Automation
             AutomationEntry entry,
             AutomationRunInfo.StartReason reason,
             string? changedEntity,
-            TaskCompletionSource? startTcs)
+            TaskCompletionSource? startTcs,
+            Event @event)
         {
             Dictionary<Type, object>? snapshot = null;
 
@@ -227,6 +239,9 @@ namespace hhnl.HomeAssistantNet.Automations.Automation
 
             object CreateEntitySnapshot(Type entityType)
             {
+                if (entityType == typeof(Event))
+                    return @event;
+
                 if (!entityType.IsAssignableTo(typeof(Entity)))
                     throw new InvalidOperationException($"Type {entityType} is not an entity");