+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.Automations/HomeAssistantConnection/HomeAssistantClient.cs
@@ -12,6 +12,7 @@ using System.Threading.Tasks;
 using hhnl.HomeAssistantNet.Automations.Automation;
 using hhnl.HomeAssistantNet.Automations.Utils;
 using hhnl.HomeAssistantNet.Shared.Configuration;
+using hhnl.HomeAssistantNet.Shared.Entities;
 using hhnl.HomeAssistantNet.Shared.HomeAssistantConnection;
 using MediatR;
 using Microsoft.Extensions.Hosting;
@@ -28,6 +29,7 @@ namespace hhnl.HomeAssistantNet.Automations.HomeAssistantConnection
         private readonly ILogger<HomeAssistantClient> _logger;
         private readonly IMediator _mediator;
         private readonly Channel<byte[]> _messagesToSend;
+        private readonly bool _publishEventNotification;
         private CancellationTokenSource? _cancellationTokenSource;
         private long _id;
         private Task? _receiveTask;
@@ -37,12 +39,14 @@ namespace hhnl.HomeAssistantNet.Automations.HomeAssistantConnection
         public HomeAssistantClient(
             ILogger<HomeAssistantClient> logger,
             IOptions<HomeAssistantConfig> haConfig,
-            IMediator mediator)
+            IMediator mediator,
+            IAutomationRegistry automationRegistry)
         {
             _logger = logger;
             _haConfig = haConfig;
             _mediator = mediator;
             _messagesToSend = Channel.CreateBounded<byte[]>(10);
+            _publishEventNotification = automationRegistry.HasAutomationsTrackingEvent;
         }
 
         public void Dispose()
@@ -241,8 +245,7 @@ namespace hhnl.HomeAssistantNet.Automations.HomeAssistantConnection
                         new
                         {
                             id = i,
-                            type = "subscribe_events",
-                            event_type = "state_changed",
+                            type = "subscribe_events"
                         });
 
                     break;
@@ -254,7 +257,7 @@ namespace hhnl.HomeAssistantNet.Automations.HomeAssistantConnection
 
                     break;
                 case "event":
-                    var apiEvent = await apiMessage.Event.ToObjectAsync<WebsocketApiEvent>();
+                    var apiEvent = await apiMessage.Event.ToObjectAsync<Event>();
 
                     if (apiEvent is null)
                         return;
@@ -266,7 +269,7 @@ namespace hhnl.HomeAssistantNet.Automations.HomeAssistantConnection
         }
 
 
-        private async Task HandleEventAsync(WebsocketApiEvent apiEvent)
+        private async Task HandleEventAsync(Event apiEvent)
         {
             switch (apiEvent.EventType)
             {
@@ -276,9 +279,15 @@ namespace hhnl.HomeAssistantNet.Automations.HomeAssistantConnection
                     if (eventData is null)
                         return;
 
+                    eventData.SourceEvent = apiEvent;
+
                     await _mediator.Publish(eventData);
                     break;
             }
+
+            // We only publish this event if anyone is listening.
+            if(_publishEventNotification)
+                await _mediator.Publish(new EventFiredNotification(apiEvent));
         }
 
         private async Task<WebsocketApiMessage> ReadNextEventAsync()
@@ -391,17 +400,6 @@ namespace hhnl.HomeAssistantNet.Automations.HomeAssistantConnection
             [JsonPropertyName("message")] public string? Message { get; set; }
         }
 
-        private class WebsocketApiEvent
-        {
-            [JsonPropertyName("time_fired")] public DateTimeOffset TimeFired { get; set; }
-
-            [JsonPropertyName("event_type")] public string EventType { get; set; }
-
-            [JsonPropertyName("origin")] public string Origin { get; set; }
-
-            [JsonPropertyName("data")] public JsonElement Data { get; set; }
-        }
-
         public class HomeAssistantCallFailedException : Exception
         {
             public HomeAssistantCallFailedException(string code, string message)
@@ -418,6 +416,19 @@ namespace hhnl.HomeAssistantNet.Automations.HomeAssistantConnection
             [JsonPropertyName("entity_id")] public string EntityId { get; set; }
 
             [JsonPropertyName("new_state")] public JsonElement NewState { get; set; }
+
+            [JsonIgnore]
+            public Event SourceEvent { get; set; }
+        }
+
+        public class EventFiredNotification : INotification
+        {
+            public EventFiredNotification(Event @event)
+            {
+                Event = @event;
+            }
+
+            public Event Event { get; set; }
         }
 #pragma warning restore 8618
     }