+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.Shared/Automation/AutomationRunInfo.cs
@@ -56,6 +56,7 @@ namespace hhnl.HomeAssistantNet.Shared.Automation
             EntityChanged,
             Manual,
             Schedule,
+            EventFired,
         }
     }
 }